% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at.R
\name{convert_unit}
\alias{convert_unit}
\alias{convert_unit_at}
\title{Modify a vector or dataframe-columns by converting units.}
\usage{
convert_unit(x, from, to)

convert_unit_at(x, .at, from, to)
}
\arguments{
\item{x}{A vector or dataframe.}

\item{from}{the unit in which the measurement was made.}

\item{to}{the unit to which the measurement is to be converted.}

\item{.at}{A character vector of names, positive numeric vector of
positions to include, or a negative numeric vector of positions to
exlude. Only those elements corresponding to \code{.at} will be modified.
If the \code{tidyselect} package is installed, you can use \code{vars()} and
the \code{tidyselect} helpers to select elements.}
}
\value{
An object with the same structure as \code{x} (vector or dataframe).
}
\description{
\code{convert_unit()} of a vector or \code{convert_unit_at()} specific columns of a
dataframe.
}
\examples{
# For all units see ?measurements::conv_unit()
convert_unit(10, "mm2", "hectare")
convert_unit(1:3, from = "m", to = "mm")
convert_unit(1:3, "mph", "kph")

dfm <- data.frame(dbh = c(10, 100), name = c(10, 100))
convert_unit_at(dfm, .at = "dbh", from = "mm2", to = "hectare")
# All columns
convert_unit_at(dfm, .at = c("dbh", "name"), from = "mm2", to = "hectare")
# Same
convert_unit_at(dfm, .at = names(dfm), from = "mm2", to = "hectare")
}
\seealso{
\code{\link[measurements:conv_unit]{measurements::conv_unit()}}, \code{\link[purrr:map_at]{purrr::map_at()}}.

Other general functions to perform common transforms: \code{\link{standardize_at}}
}
\concept{general functions to perform common transforms}
