% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.misc-package.R
\docType{package}
\name{fgeo.misc-package}
\alias{fgeo.misc}
\alias{fgeo.misc-package}
\title{fgeo.misc: Miscellaneous Functions}
\description{
Miscellaneous functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.misc}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.misc/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder]
}

}
\keyword{internal}
