% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_top.R
\name{in_top}
\alias{in_top}
\alias{top}
\title{Pick the top \code{n} values from the head or tail of a vector.}
\usage{
in_top(x, n = 1L)

top(x, n = 1L)
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{A single integer. Doubles are silently coerced to integer.}
}
\value{
\itemize{
\item \code{top()} returns a vector of the same type as the input.
\item \code{in_top()} returns a logical vector.
}
}
\description{
Pick the top \code{n} values from the head or tail of a vector.
}
\examples{

# Pick from the head
top(1:3)
# Same
top(1:3, 1L)

# Pick from the tail
top(1:3, -2L)

# Useful for filtering
in_top(1:3)
in_top(1:3, -1L)

subset(mtcars, in_top(cyl))

subset(mtcars, in_top(cyl) & in_top(carb))

subset(mtcars, in_top(cyl, -2L))

# Careful: Remember that `FALSE` evaluates to 0
lgl <- c(TRUE, FALSE, TRUE)
int <- as.integer(lgl)
int

top(lgl)
top(int)

top(lgl, -1)
top(int, -1)

}
\seealso{
Other general functions to pick or drop rows of a dataframe: \code{\link{pick_recensus}},
  \code{\link{pick_top}}
}
\concept{general functions to pick or drop rows of a dataframe}
