% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout_xlff_to_output.R
\name{misc_example}
\alias{misc_example}
\title{Path to directory containing example data.}
\usage{
misc_example(path)
}
\arguments{
\item{path}{Path to a file (with extension) from inst/extdata/.}
}
\value{
String (a path)
}
\description{
Path to directory containing example data.
}
\examples{
misc_example("xl")
dir(misc_example("xl"))
}
