% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_plotname.R
\name{pick_plotname}
\alias{pick_plotname}
\title{Pick the given or first (alpha sorted) plot from a ViewFullTable.}
\usage{
pick_plotname(vft, plot_nm = NULL)
}
\arguments{
\item{vft}{Dataframe; particularly a ForestGEO ViewFullTable.}

\item{plot_nm}{Length-1 character vector of the value of \code{PlotName} (or
\code{plotname}) to pick from \code{vft}.}
}
\value{
A dataframe.
}
\description{
Pick the given or first (alpha sorted) plot from a ViewFullTable.
}
\examples{
vft <- data.frame(PlotName = c("a", "b"), stringsAsFactors = FALSE)
pick_plotname(vft)
pick_plotname(vft, "b")

}
\seealso{
Other functions to pick or drop rows of a ForestGEO dataframe: \code{\link{drop_twice_dead}}
}
\concept{functions for fgeo vft}
\concept{functions to pick or drop rows of a ForestGEO dataframe}
