% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsheets_list.R
\name{xl_list}
\alias{xl_list}
\alias{xlbooks_list}
\title{Import excel files from a directory into a list.}
\usage{
xl_list(path_dir, ...)

xlbooks_list(path_dir, ...)
}
\arguments{
\item{path_dir}{String; the path to a directory containing the files to read
(all must be of appropriate format; see examples).}

\item{...}{Arguments passed to the reader function.}
}
\value{
A list of dataframes.
}
\description{
These functions read from a specific directory where every file has its own extension
as indicated by each function's name. Notice that function names have the format
input_output, i.e. file-extension_list. If none of these functions do what
you want, create your own with \code{\link[=read_with]{read_with()}}.
}
\examples{
path_xl <- misc_example("xl")
path_xl
dir(path_xl)
xl_list(path_xl)

path_books <- misc_example("multiple_workbooks")
dir(path_books)
xlbooks_list(path_books)
}
\seealso{
Other general functions to import data: \code{\link{read_with}}
}
