% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsheets_list.R
\name{xl_list}
\alias{xl_list}
\title{Import excel files from a directory into a list.}
\usage{
xl_list(path)

xl_list(path)
}
\arguments{
\item{path}{String; the path to a directory containing the files to read
(all must be of appropriate format; see examples).}

\item{...}{Arguments passed to the reader function.}
}
\description{
Import excel files from a directory into a list.
}
\examples{
path_xl <- misc_example("xl")
path_xl
dir(path_xl)
xl_list(path_xl)
}
